plugin/pyproject.toml
[project]
name = "kubectl-marimo"
version = "0.1.0"
description = "kubectl plugin for deploying marimo notebooks to Kubernetes"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "kubernetes>=28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
kubectl-marimo = "kubectl_marimo.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kubectl_marimo"]
